+++ lib/widgets/image_page.dart
@@ -32,7 +32,6 @@ library;
 import 'dart:async';
 import 'dart:io';
 import 'dart:math' as math;
-import 'dart:ui' as ui;
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
@@ -43,16 +42,14 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:gap/gap.dart';
 import 'package:markdown_tooltip/markdown_tooltip.dart';
-import 'package:pdf/widgets.dart' as pw;
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 import 'package:rattle/constants/sunken_box_decoration.dart';
 import 'package:rattle/constants/temp_dir.dart';
 import 'package:rattle/providers/settings.dart';
-import 'package:rattle/utils/select_file.dart';
+import 'package:rattle/utils/save_plot.dart';
 import 'package:rattle/utils/show_image_dialog.dart';
-import 'package:rattle/utils/show_ok.dart';
 
 /// The [path] is required. It is the path to a SVG or PNG file to display in
 /// the app and to open/save externally. If [display] is also provided then this
@@ -113,65 +110,6 @@ class ImagePage extends ConsumerWidget {
 
       return null;
     }
-  }
-
-  // Convert the file [svgPath] return [Future] image bytes in PNG format.
-  //
-  // Throws an [Exception] if the conversion fails.
-
-  Future<ByteData> _svgToImageBytes(String svgPath) async {
-    final svgString = await File(svgPath).readAsString();
-
-    final pictureInfo = await vg.loadPicture(SvgStringLoader(svgString), null);
-
-    final recorder = ui.PictureRecorder();
-    final canvas = ui.Canvas(recorder);
-    final size = pictureInfo.size;
-
-    canvas.scale(1.0, 1.0);
-
-    pictureInfo.picture.toImage(size.width.toInt(), size.height.toInt());
-
-    final image = await pictureInfo.picture.toImage(
-      size.width.toInt(),
-      size.height.toInt(),
-    );
-    final byteData = await image.toByteData(format: ui.ImageByteFormat.png);
-
-    if (byteData == null) {
-      throw Exception('Failed to convert SVG to image bytes');
-    }
-
-    return byteData;
-  }
-
-  // Export the SVG file [svgPath] into a PDF file [pdfPath].
-
-  Future<void> _exportToPdf(String svgPath, String pdfPath) async {
-    final pngBytes = await _svgToImageBytes(svgPath);
-
-    final pdf = pw.Document();
-    pdf.addPage(
-      pw.Page(
-        build: (pw.Context context) {
-          return pw.Center(
-            child: pw.Image(pw.MemoryImage(pngBytes.buffer.asUint8List())),
-          );
-        },
-      ),
-    );
-
-    final file = File(pdfPath);
-    await file.writeAsBytes(await pdf.save());
-  }
-
-  // Export the SVG file [svgPath] into a PNG file [pngPath].
-
-  Future<void> _exportToPng(String svgPath, String pngPath) async {
-    final pngBytes = await _svgToImageBytes(svgPath);
-
-    final file = File(pngPath);
-    await file.writeAsBytes(pngBytes.buffer.asUint8List());
   }
 
   @override
@@ -346,49 +284,7 @@ class ImagePage extends ConsumerWidget {
                         ''',
                       child: IconButton(
                         icon: const Icon(Icons.save, color: Colors.blue),
-                        onPressed: () async {
-                          String fileName = path.split('/').last;
-                          String? pathToSave = await selectFile(
-                            defaultFileName: fileName,
-                            allowedExtensions: ['svg', 'pdf', 'png'],
-                          );
-                          if (pathToSave != null) {
-                            String extension =
-                                pathToSave.split('.').last.toLowerCase();
-                            if (extension == 'svg') {
-                              await File(path).copy(pathToSave);
-                            } else if (extension == 'pdf') {
-                              await _exportToPdf(path, pathToSave);
-                            } else if (extension == 'png') {
-                              if (path.toLowerCase().endsWith('.svg')) {
-                                await _exportToPng(path, pathToSave);
-                              } else {
-                                // If source is already PNG, just copy it.
-
-                                await File(path).copy(pathToSave);
-                              }
-                            } else if (context.mounted) {
-                              // If the user selected an unsupported file
-                              // extension show an error dialog.
-
-                              showOk(
-                                title: 'Error',
-                                context: context,
-                                content: //const Text(
-                                    '''
-
-                                      An unsupported filename extension was
-                                      provided: .$extension.  Please try again
-                                      and select a filename with one of the
-                                      supported extensions: .svg, .pdf, or .png.
-
-                                      ''',
-                              );
-                            } else {
-                              return;
-                            }
-                          }
-                        },
+                        onPressed: () => savePlot(context, path),
                       ),
                     ),
                     const Gap(5),